% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prokounter.R
\name{.getRecovAbPat}
\alias{.getRecovAbPat}
\title{Get recovered abundance dependent statistics for a given genus in each sample in the data.}
\usage{
.getRecovAbPat(mat, genus, g, adt = TRUE)
}
\arguments{
\item{mat}{a taxa x samples matrix.}

\item{genus}{a vector indicating the genus category.}

\item{g}{the genus for which the calculations are to be done.}

\item{div.too}{if}
}
\value{
A dataframe with components
        \itemize{
        \item{lr - log abundances of the genus in each sample}
        \item{lm - log number of taxa for the genus in each sample}
        \item{samples - the sample name to which the row-wise entries correspond to}
        \item{g - the genus name given}
        \item{chao1 - the chao1 estimator as calculated from the package Vegan }
        \item{chao1.se - the standard error fot the chao1 estimate }
        \item{ace - the ACE estimator as calculated from the package Vegan }
        \item{ace.se - the standard error for the ACE estimate}
        }
}
\description{
Get recovered abundance dependent statistics for a given genus in each sample in the data.
}
