% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prokounter.R
\name{.plot.fr}
\alias{.plot.fr}
\title{Plots the fitted \code{fR} trend from the manuscript.}
\usage{
.plot.fr(pkobj)
}
\arguments{
\item{pkobj}{the output object of the getProkounterTrends() function.}
}
\description{
Plots the fitted \code{fR} trend from the manuscript.
}
