% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prokounter.R
\name{getDRforGenera}
\alias{getDRforGenera}
\title{Genus-specific differential richness inference for 16S surveys}
\usage{
getDRforGenera(
  pkobj,
  des,
  resp.type = c("lm", "chao1", "ace"),
  glm.fam = "poisson",
  depth.inc = FALSE,
  ...
)
}
\arguments{
\item{pkobj}{the output object from prokounter::getProkounterTrends function.}

\item{des}{the design matrix.}

\item{resp.type}{Three possible options. By default, \code{"lm"} when the association test is to be done with respect to observed number of taxa (sampled richness).
\code{"chao1"} when the association test is to be done with an asymptotic richness estimator Chao1.
\code{"ace"} when the association test is to be done with an asymptotic richness estimator ACE.
In 16S surveys, we have found these options to be equivalent because sampled richness is heavily correlated with asymptotic richness estimates.}

\item{glm.fam}{the glm family. Default "poisson" when estimation is performed by \code{glm}.
"NB" for negative binomial, when eestimation is performed with \code{MASS::glm.nb}.
#' @param depth.inc TRUE if logged sample depth (a sample's total count) needs be incorporated as a predictor. Default FALSE as we find prokounter's trend \code{ft} to capture sampling effort well.}

\item{depth.inc}{TRUE if logged sample depth (a sample's total count) needs be incorporated as a predictor. Default FALSE as we find prokounter's trend \code{ft} to capture sampling effort well.}

\item{...}{other parameters to pass to \code{glm} or \code{glm.nb}}
}
\value{
A list with two entries.
\itemize{
  \item{ fits - holds the glm fits for each genus.  }
  \item{ tab - a list where each entry holds the association statistics (effect size, standard error, test statistic, p-value)
        corresponding to each column in the design matrix, row-wise for each genus.}
}
}
\description{
Obtains differential richness for each genus in the study and a given design matrix.
}
\examples{
#Obtain counts matrix and sample group information
require(metagenomeSeq)
require(prokounter)
data(mouseData)
taxaBySamplesCountMatrix <- MRcounts(mouseData, norm=FALSE)
des <- model.matrix( ~diet,  pData(mouseData) ) #the rownames of the design matrix must contain corresponding sample names.
genus <- as.character(fData(mouseData)$genus)
pkt <- prokounter::getProkounterTrends( mat = taxaBySamplesCountMatrix,genus=genus)
dr.gw <- getDRforGenera( pkobj=pkt, des = des )
dr.gw$tab
}
\seealso{
\code{\link{getDRforTaxaCollection}}, \code{\link{getDRforSampleGroups}}
}
