% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prokounter.R
\name{getDRforSampleGroups.bootWholeData}
\alias{getDRforSampleGroups.bootWholeData}
\title{Bootstrap t confidence intervals for sample-wide differential richness inference form Prokounter.}
\usage{
getDRforSampleGroups.bootWholeData(
  esetobj,
  des,
  genusLabel = "Genus",
  ncore = 4,
  B = 5000,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{esetobj}{the ExpressionSet object}

\item{des}{design matrix for sample-wide differential richness inference}

\item{genusLabel}{\code{fData(esetobj)[[genusLabel]]} should return the genus categories. More generally, this can be any category within which we can assume most features are false.}

\item{ncore}{number of cores or processes to use for bootstrap}

\item{B}{Number of bootstrap samples. We strongly recommend experimenting with this parameter for your dataset.}

\item{alpha}{type I error rate. 1-alpha bootstrap t confidence intervals are constructed.}

\item{....}{other parameters to pass to \code{\link{getDRforSampleGroups}}.}
}
\value{
A list where each entry holds the results for regression coefficients corresponding to each column in the design matrix.
\itemize{
  \item{ fit - holds the base result from \code{\link{getDRforSampleGroups}}   }
  \item{ bootstrapt.ci - the bootstrap confidence interval at the prescribed confidence level. }
}
}
\description{
Developmental. The procedure aims to allow gauging reproducibility of inferences over the fitted prokounter trends (\code{ft} in manuscript) in \code{\link{getDRforSampleGroups}}.
We strongly recommend experimenting with the number of bootstrap samples (parameter B), until the returned confidence intervals stabilize.
We have found this to vary with data set size and complexities.
The code can be adapted for genus-level richness inference.
The procedure needs an ExpressionSet object (ref. example). This allows us to handle genus level calculations in an easier fashion in bootstrap samples.
}
\examples{
 #Obtain counts matrix and sample group information
library(metagenomeSeq)
library(prokounter)
data(mouseData)
taxaBySamplesCountMatrix <- MRcounts(mouseData, norm=FALSE)
des <- model.matrix( ~diet,  pData(mouseData) ) #the rownames of the design matrix must contain corresponding sample names.
mouseEsetObj <- ExpressionSet( assayData = taxaBySamplesCountMatrix, phenoData = phenoData(mouseData), featureData = featureData(mouseData) )
genusLabel <- "genus" #fData(mouseEsetObj)[[genusLabel]] yields the genus annotations.
dr.sw.boot <- getDRforSampleGroups.bootWholeData( esetobj =  mouseEsetObj, des=des, genusLabel="genus", B = 5000 )
dr.sw.boot
}
\seealso{
\code{\link{getDRforGenera}}, \code{\link{getDRforSampleGroups}}
}
