% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prokounter.R
\name{getDRforSampleGroups}
\alias{getDRforSampleGroups}
\title{Sample-wide differential richness inference for 16S surveys}
\usage{
getDRforSampleGroups(
  pkobj,
  des,
  resp.type = c("lm", "chao1", "ace"),
  glm.fam = "NB",
  depth.inc = FALSE,
  fg.inc = TRUE,
  fg.sigs.only = TRUE,
  ...
)
}
\arguments{
\item{pkobj}{the output object from \code{\link{getProkounterTrends}} function.}

\item{des}{the design matrix.}

\item{resp.type}{Three possible options. By default, \code{"lm"} when the association test is to be done with respect to observed number of taxa (sampled richness).
\code{"chao1"} or \code{"ace"} when the association test is to be done with the respective asymptotic richness estimators.
In 16S surveys, we often find these options to be equivalent.}

\item{glm.fam}{the family function for the generalized linear model. \code{"poisson"} when estimation is performed by \code{glm}.
Default \code{"NB"} for negative binomial, when estimation is performed with \code{MASS::glm.nb}.}

\item{depth.inc}{TRUE if logged sample depth (a sample's total count) needs be incorporated as a predictor. Default FALSE as we find prokounter's trend \code{ft} to capture sampling effort well.}

\item{fg.inc}{TRUE by default. TRUE if prokounter genus specific effects (the fg term in the prokounter trend model) need to be incorporated in deriving the net technical effect.
This is available only when the prokounter trend was derived with \code{fit.proc="ss"} option from the function \code{\link{getProkounterTrends}}.}

\item{fg.sigs.only}{TRUE if only significant genus effects (fg term in the prokounter trend model) needs to be incorporated.
The significance of the effects is calculated using the Benjamini-Hoschberg procedure at a default false discovery rate threshold of .1.
This FDR threshold can be controlled when running \code{\link{getProkounterTrends}} through the option \code{fg.fdr}.}

\item{...}{other parameters to pass to \code{glm} or \code{glm.nb}}
}
\value{
A list with two entries.
\itemize{
  \item{ \code{fit} - holds the GLM fit  }
  \item{ \code{tab} - the association statistics (effect size, standard error, test statistic, p-value)
        corresponding to each column in the design matrix, row-wise for each sample-group.}
}
}
\description{
Obtains differential richness for sample groups and a given design matrix.
}
\examples{
#Obtain counts matrix and sample group information
library(metagenomeSeq)
library(prokounter)
data(mouseData)
taxaBySamplesCountMatrix <- MRcounts(mouseData, norm=FALSE)
des <- model.matrix( ~diet,  pData(mouseData) ) #the rownames of the design matrix must contain corresponding sample names.
genus <- as.character(fData(mouseData)$genus)
pkt <- prokounter::getProkounterTrends( mat = taxaBySamplesCountMatrix,genus=genus)
dr.sw <- getDRforSampleGroups( pkobj=pkt, des = des )
dr.sw$tab
}
\seealso{
\code{\link{getDRforTaxaCollection}}, \code{\link{getDRforSampleGroups}}
}
