% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prokounter.R
\name{getDRforTaxaCollection}
\alias{getDRforTaxaCollection}
\title{Taxa collection-specific differential richness inference for 16S surveys.}
\usage{
getDRforTaxaCollection(
  pkobj,
  des,
  genus2collection = NULL,
  resp.type = c("lm", "chao1", "ace"),
  glm.fam = "poisson",
  depth.inc = FALSE,
  fg.inc = TRUE,
  fg.sigs.only = TRUE,
  ...
)
}
\arguments{
\item{pkobj}{the output object from prokounter::getProkounterTrends function.}

\item{des}{the design matrix.}

\item{genus2collection}{a data frame or a character matrix with the first column containing the genus
identifier and the second column the collection it has been asssigned to.
The genus identifier must be the same as those found in the original data for which the
prokounter::getProkounterTrends trends were established.}

\item{resp.type}{Three possible options. By default, \code{"lm"} when the association test is to be done with respect to observed number of taxa (sampled richness).
\code{"chao1"} when the association test is to be done with an asymptotic richness estimator Chao1.
\code{"ace"} when the association test is to be done with an asymptotic richness estimator ACE.
In 16S surveys, we have found these options to be equivalent because sampled richness is heavily correlated with asymptotic richness estimates.}

\item{glm.fam}{the glm family. Default "poisson" when estimation is performed by \code{glm}.
"NB" for negative binomial, when eestimation is performed with \code{MASS::glm.nb}.}

\item{depth.inc}{TRUE if logged sample depth (a sample's total count) needs be incorporated as a predictor. Default FALSE as we find prokounter's trend \code{ft} to capture sampling effort well.}

\item{fg.inc}{TRUE by default. TRUE if prokounter genus specific effects (the fg term in the prokounter trend model) need to be incorporated in deriving the net technical effect.
This is available only when the prokounter trend was derived with \code{fit.proc="ss"} option from the function \code{\link{getProkounterTrends}}.}

\item{fg.sigs.only}{TRUE if only significant genus effects (fg term in the prokounter trend model) needs to be incorporated.
The significance of the effects is calculated using the Benjamini-Hoschberg procedure at a default false discovery rate threshold of .1.
This FDR threshold can be controlled when running \code{\link{getProkounterTrends}} through the option \code{fg.fdr}.}

\item{...}{other parameters to pass to \code{glm} or \code{glm.nb}}
}
\value{
A list where each entry holds the association statistics (effect size, standard error, test statistic, p-value) corresponding to each column in the design matrix.
\itemize{
  \item{ fits - holds the glm fits for each collection  }
  \item{ tab - a list where each entry holds the association statistics (effect size, standard error, test statistic, p-value)
        corresponding to each column in the design matrix, row-wise for each collection.}
}
}
\description{
Obtains differential richness for collections of taxa and a given design matrix.
Each collection is viewed as a set of one or multiple genera. The function \code{\link{getDRforGenera}} is a special case of this function.
}
\examples{
#Obtain counts matrix and sample group information
require(metagenomeSeq)
require(prokounter)
data(mouseData)
taxaBySamplesCountMatrix <- MRcounts(mouseData, norm=FALSE)
des <- model.matrix( ~diet,  pData(mouseData) ) #the rownames of the design matrix must contain corresponding sample names.
genus <- as.character(fData(mouseData)$genus)
genus2collection <- fData(mouseData)[,c("genus", "phylum")]
pkt <- prokounter::getProkounterTrends( mat = taxaBySamplesCountMatrix,genus=genus)
dr.cw <- getDRforTaxaCollection( pkobj=pkt, des = des, genus2collection=genus2collection )
dr.cw$tab

}
\seealso{
\code{\link{getDRforGenera}}, \code{\link{getDRforSampleGroups}}
}
