% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prokounter.R
\name{getProkounterTrends}
\alias{getProkounterTrends}
\title{Obtains the prokounter within-genus taxa accumulation trend.}
\usage{
getProkounterTrends(
  mat,
  genus,
  fit.proc = "ss",
  toig = c("g__", "unknown", "NA", "", "undef"),
  fg.fdr = 0.1,
  plt = FALSE,
  kn.rand = FALSE,
  ...
)
}
\arguments{
\item{mat}{a taxa x samples count matrix obtained from a 16S survey.
We recommend not to post-process or filter 16S data
so as to capture the accumulation rates of error-like low abundant, sparse taxa.}

\item{genus}{a character vector indicating the genus category for each taxa (i.e., row).
Total length must equal the number of rows in the data matrix \code{mat}. More generally, this can be any annotation category within which one can assume most taxa discoveries are false.}

\item{fit.proc}{Trend deriving procedure. "ss" when the trend is estimated as smoothing spline (default). "lo" when a loess smoother is used.}

\item{toig}{Genera categories to ignore. By default this includes unknown or undefined categories \code{c("g__", "unknown", "NA", "", "undef")}.}

\item{fg.fdr}{False discovery rate threshold for assessing the significance of \code{fG} effects.
By default set to \code{.1}. This parameter determines the significant genus effects incorporated in the technical trend while
assessing differntial richness in functions \code{\link{getDRforSampleGroups}}, \code{\link{getDRforGenera}}, and \code{\link{getDRforTaxaCollection}}.}

\item{plt}{By default, set to FALSE. TRUE if recovered abundance trend \code{fR} term in manuscript's notation needs to be plotted.
If using \code{fit.proc="ss"}, the random choice of knots may affect the shape of \code{fR}, but the net predicted mean of the linear predictor will be similar.}

\item{kn.rand}{FALSE by default. TRUE if ties in the choice of spline knots are broken at random.}

\item{...}{other parameters to control the fits of smoothing spline \code{gss::ssanova} or the loess (loess) fitting procedures.}
}
\value{
A list with three components.
       \itemize{
       \item{fit - the fitted trends object.}
       \item{
       df - a data frame with multiple components
       \itemize{
        \item{lr - log recovered abundances of the genus across samples}
        \item{lm - log number of taxa for the genus across samples}
        \item{samples - the sample name to which the row-wise entries correspond to}
        \item{g - the genus name given}
        \item{fit.proc - fitting procedure used}
        \item{chao1 - the chao1 estimator as calculated from the package Vegan }
        \item{chao1.se - the standard error fot the chao1 estimate }
        \item{ace - the ACE estimator as calculated from the package Vegan }
        \item{ace.se - the standard error for the ACE estimate}
        \item{ltv - the marginal recovered abundance dependent component (\code{fR} term in the main manuscript) with an additive constant}
        \item{ltv.se - the standard error of ltv }
        \item{gtv - the genus effect term. \code{fG} in the manuscript}
        \item{gtv.se - the standard error of \code{fG}}
        \item{gltv - estimates \code{fR + fG} with an additive constant  }
        \item{gltv.se - standard error of gltv }
        \item{gltv.wonly.gsigs - \code{fR + fG} with an additive constant, but only the statistically significant \code{fG} are incorporated in the sum.  }
        }
       }
       \item{ug - the set of genera over which the calculations were done.  }
       }
}
\description{
The result objects generated with this function is used for differential richness inference
sample-wide (through the function \code{\link{getDRforSampleGroups}}),
genus-wide (\code{\link{getDRforGenera}}), and for collections of genera (e.g., phyla, \code{\link{getDRforTaxaCollection}}).
}
